+++ Model/Api.php
@@ -26,6 +26,8 @@ use Magento\Framework\Cache\InvalidateLogger;
 use Fastly\Cdn\Helper\Data;
 use Fastly\Cdn\Helper\Vcl;
 use Psr\Log\LoggerInterface;
+use Magento\Backend\Model\Auth\Session\Proxy;
+use Magento\Framework\App\State;
 
 /**
  * Class Api
@@ -77,13 +79,26 @@ class Api
     private $vcl;
 
     /**
+     * @var Proxy
+     */
+    private $authSession;
+
+    /**
+     * @var State
+     */
+    private $state;
+
+    /**
      * Api constructor.
+     *
      * @param Config $config
      * @param CurlFactory $curlFactory
      * @param InvalidateLogger $logger
      * @param Data $helper
      * @param LoggerInterface $log
      * @param Vcl $vcl
+     * @param Proxy $authSession
+     * @param State $state
      */
     public function __construct(
         Config $config,
@@ -91,7 +106,9 @@ class Api
         InvalidateLogger $logger,
         Data $helper,
         LoggerInterface $log,
-        Vcl $vcl
+        Vcl $vcl,
+        Proxy $authSession,
+        State $state
     ) {
         $this->config = $config;
         $this->curlFactory = $curlFactory;
@@ -99,6 +116,8 @@ class Api
         $this->helper = $helper;
         $this->log = $log;
         $this->vcl = $vcl;
+        $this->authSession = $authSession;
+        $this->state = $state;
     }
 
     /**
@@ -157,6 +176,7 @@ class Api
      */
     public function cleanBySurrogateKey($keys)
     {
+        $type = 'clean by key on ';
         $uri = $this->_getApiServiceUri() . 'purge';
         $num = count($keys);
         $result = false;
@@ -172,15 +192,27 @@ class Api
 
         foreach ($collection as $keys) {
             $payload = json_encode(['surrogate_keys' => $keys]);
-            if ($result = $this->_purge($uri, \Zend_Http_Client::POST, $payload)) {
+            if ($result = $this->_purge($uri, null, \Zend_Http_Client::POST, $payload)) {
                 foreach ($keys as $key) {
                     $this->logger->execute('surrogate key: ' . $key);
                 }
             }
 
-            if ($this->config->areWebHooksEnabled() && $this->config->canPublishKeyUrlChanges()) {
+            $canPublishKeyUrlChanges = $this->config->canPublishKeyUrlChanges();
+            $canPublishPurgeChanges = $this->config->canPublishPurgeChanges();
+
+            if ($this->config->areWebHooksEnabled() && ($canPublishKeyUrlChanges || $canPublishPurgeChanges)) {
                 $status = $result ? '' : 'FAILED ';
                 $this->sendWebHook($status . '*clean by key on ' . join(" ", $keys) . '*');
+
+                $canPublishPurgeByKeyDebugBacktrace = $this->config->canPublishPurgeByKeyDebugBacktrace();
+                $canPublishPurgeDebugBacktrace = $this->config->canPublishPurgeDebugBacktrace();
+
+                if ($canPublishPurgeByKeyDebugBacktrace == false && $canPublishPurgeDebugBacktrace == false) {
+                    return $result;
+                }
+
+                $this->stackTrace($type . join(" ", $keys));
             }
         }
 
@@ -201,25 +233,26 @@ class Api
         }
         $this->purged = true;
 
+        $type = 'clean/purge all';
         $uri = $this->_getApiServiceUri() . 'purge_all';
-        if ($result = $this->_purge($uri)) {
+        if ($result = $this->_purge($uri, null)) {
             $this->logger->execute('clean all items');
         }
 
-        if ($this->config->areWebHooksEnabled() && $this->config->canPublishPurgeAllChanges()) {
+        $canPublishPurgeAllChanges = $this->config->canPublishPurgeAllChanges();
+        $canPublishPurgeChanges = $this->config->canPublishPurgeChanges();
+
+        if ($this->config->areWebHooksEnabled() && ($canPublishPurgeAllChanges || $canPublishPurgeChanges)) {
             $this->sendWebHook('*initiated clean/purge all*');
 
-            if ($this->config->canPublishDebugBacktrace() == false) {
+            $canPublishPurgeAllDebugBacktrace = $this->config->canPublishPurgeAllDebugBacktrace();
+            $canPublishPurgeDebugBacktrace = $this->config->canPublishPurgeDebugBacktrace();
+
+            if ($canPublishPurgeAllDebugBacktrace == false && $canPublishPurgeDebugBacktrace == false) {
                 return $result;
             }
 
-            $stackTrace = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS);
-            $trace = [];
-            foreach ($stackTrace as $row => $data) {
-                $trace[] = "#{$row} {$data['file']}:{$data['line']} -> {$data['function']}()";
-            }
-
-            $this->sendWebHook('*Purge all backtrace:*```' .  implode("\n", $trace) . '```');
+            $this->stackTrace($type);
         }
 
         return $result;
@@ -229,12 +262,13 @@ class Api
      * Send purge request via Fastly API
      *
      * @param $uri
+     * @param $type
      * @param string $method
      * @param null $payload
      * @return bool
      * @throws \Zend_Uri_Exception
      */
-    private function _purge($uri, $method = \Zend_Http_Client::POST, $payload = null)
+    private function _purge($uri, $type, $method = \Zend_Http_Client::POST, $payload = null)
     {
 
         if ($method == 'PURGE') {
@@ -264,6 +298,7 @@ class Api
             );
         }
 
+        $result = true;
         try {
             $client = $this->curlFactory->create();
             $client->setConfig(['timeout' => self::PURGE_TIMEOUT]);
@@ -281,10 +316,24 @@ class Api
             }
         } catch (\Exception $e) {
             $this->logger->critical($e->getMessage(), $uri);
-            return false;
+            $result = false;
         }
 
-        return true;
+        if (empty($type)) {
+            return $result;
+        }
+
+        if ($this->config->areWebHooksEnabled() && $this->config->canPublishPurgeChanges()) {
+            $this->sendWebHook('*initiated ' . $type .'*');
+
+            if ($this->config->canPublishPurgeDebugBacktrace() == false) {
+                return $result;
+            }
+
+            $this->stackTrace($type);
+        }
+
+        return $result;
     }
 
     /**
@@ -743,7 +792,15 @@ class Api
     {
         $url = $this->config->getIncomingWebhookURL();
         $messagePrefix = $this->config->getWebhookMessagePrefix();
-        $currentUsername = 'System'; // TO DO: Fetch current admin username
+        $currentUsername = 'System';
+        try {
+            if ($this->state->getAreaCode() == 'adminhtml') {
+                $currentUsername = $this->authSession->getUser()->getUserName();
+            }
+        } catch (\Exception $e) {
+            $this->log->log(100, 'Failed to retrieve Area Code');
+        }
+
         $storeName = $this->helper->getStoreName();
         $storeUrl = $this->helper->getStoreUrl();
 
@@ -1164,5 +1221,16 @@ class Api
         }
 
         return json_decode($responseBody);
+    }
+
+    private function stackTrace($type)
+    {
+        $stackTrace = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS);
+        $trace = [];
+        foreach ($stackTrace as $row => $data) {
+            $trace[] = "#{$row} {$data['file']}:{$data['line']} -> {$data['function']}()";
+        }
+
+        $this->sendWebHook('*'. $type .' backtrace:*```' .  implode("\n", $trace) . '```');
     }
 }