+++ Model/Product/Image.php
@@ -302,7 +302,10 @@ class Image extends ImageModel
         $url = $this->getBaseFileUrl($baseFile);
 
         $this->fastlyParameters['quality'] = $this->_quality;
-        $this->fastlyParameters['bg-color'] = implode(',', $this->_backgroundColor);
+
+        if ($this->_scopeConfig->isSetFlag(Config::XML_FASTLY_IMAGE_OPTIMIZATION_BG_COLOR) == true) {
+            $this->fastlyParameters['bg-color'] = implode(',', $this->_backgroundColor);
+        }
         if ($this->_keepAspectRatio == true) {
             $this->fastlyParameters['fit'] = 'bounds';
         }
@@ -311,7 +314,7 @@ class Image extends ImageModel
 
     public function getBaseFileUrl($baseFile)
     {
-        if ($baseFile === null) {
+        if ($baseFile === null || $this->isBaseFilePlaceholder()) {
             $url = $this->_assetRepo->getUrl(
                 "Magento_Catalog::images/product/placeholder/{$this->getDestinationSubdir()}.jpg"
             );