+++ view/adminhtml/web/js/uploadvcl.js
@@ -366,8 +366,8 @@ define([
         });
 
         $('body').on('click', 'button.fastly-delete-snippet-icon', function () {
-            var snippet_id = $(this).data('snippet-id');
-            var closestTr = $(this).closest('tr');
+            let snippet_id = $(this).data('snippet-id');
+            let closestTr = $(this).closest('tr');
             if (confirm("Are you sure you want to delete "+ snippet_id +"?")) {
                 vcl.deleteCustomSnippet(snippet_id, true).done(function (response) {
                     if (response.status == true) {
@@ -378,6 +378,54 @@ define([
                     vcl.showErrorMessage($.mage.__('An error occurred while processing your request. Please try again.'));
                 });
             }
+        });
+
+        $('body').on('click', 'button.fastly-edit-snippet-icon', function () {
+            let snippet_id = $(this).data('snippet-id');
+            if (isAlreadyConfigured !== true) {
+                $(this).attr('disabled', true);
+                return alert($.mage.__('Please save config prior to continuing.'));
+            }
+
+            vcl.resetAllMessages();
+
+            $.when(
+                $.ajax({
+                    type: "GET",
+                    url: config.serviceInfoUrl,
+                    showLoader: true
+                })
+            ).done(function (service) {
+
+                if (service.status === false) {
+                    return errorCustomSnippetBtnMsg.text($.mage.__('Please check your Service ID and API token and try again.')).show();
+                }
+
+                active_version = service.active_version;
+                next_version = service.next_version;
+                service_name = service.service.name;
+                $.when(
+                    $.ajax({
+                        type: "GET",
+                        url: config.getCustomSnippet,
+                        data: {'snippet_id': snippet_id},
+                        showLoader: true,
+                    })
+                ).done(function (response) {
+                    if (response.status === true) {
+                        vcl.showPopup('fastly-custom-snippet-edit-options');
+                        $('.upload-button span').text('Save');
+                        $('#custom_snippet_name').val(response.name);
+                        $('#original_snippet_name').val(response.original);
+                        $('#custom_snippet_type').val(response.type);
+                        $('#custom_snippet_priority').val(response.priority);
+                        $('#custom_snippet_content').val(response.content);
+                    }
+                });
+
+            }).fail(function () {
+                return errorCustomSnippetBtnMsg.text($.mage.__('An error occurred while processing your request. Please try again.')).show();
+            });
         });
 
         // delete acl container button
@@ -629,6 +677,7 @@ define([
                 next_version = service.next_version;
                 service_name = service.service.name;
                 vcl.showPopup('fastly-custom-snippet-options');
+                $('.upload-button span').text('Create');
 
             }).fail(function () {
                 return errorCustomSnippetBtnMsg.text($.mage.__('An error occurred while processing your request. Please try again.')).show();
@@ -1259,7 +1308,8 @@ define([
                 $.each(snippets, function (index, snippet) {
                     html += "<tr id='fastly_" + index + "'>";
                     html += "<td><input data-snippetId='"+ index + "' id='snippet_" + index + "' value='"+ snippet +"' disabled='disabled' class='input-text' type='text'></td>";
-                    html += "<td class='col-actions'><button class='action-delete fastly-delete-snippet-icon' data-snippet-id='" + snippet + "' id='fastly-delete-snippet"+ index + "' title='Delete custom snippet' type='button'></td></tr>";
+                    html += "<td class='col-actions'><button class='action-delete fastly-edit-snippet-icon' data-snippet-id='" + snippet + "' id='fastly-edit-snippet"+ index + "' title='Edit custom snippet' type='button'></button>";
+                    html += "<span>&nbsp;&nbsp;</span><button class='action-delete fastly-delete-snippet-icon' data-snippet-id='" + snippet + "' id='fastly-delete-snippet"+ index + "' title='Delete custom snippet' type='button'></button></td></tr>"
                 });
                 if (html != '') {
                     $('.no-snippets').hide();
@@ -1401,6 +1451,18 @@ define([
                 });
             },
 
+            editCustomSnippet: function (snippet_id, loaderVisibility) {
+                return $.ajax({
+                    type: "GET",
+                    url: config.editCustomSnippet,
+                    showLoader: loaderVisibility,
+                    data: {'snippet_id': snippet_id},
+                    beforeSend: function (xhr) {
+                        vcl.resetAllMessages();
+                    }
+                });
+            },
+
             // Delete Acl entry item
             deleteAclItem: function (acl_id, acl_item_id, loaderVisibility) {
                 return $.ajax({
@@ -1494,11 +1556,11 @@ define([
             },
             // custom snippet creation
             setCustomSnippet: function () {
-                var custom_name = $('#custom_snippet_name').val();
-                var custom_type = $('#custom_snippet_type').val();
-                var custom_priority = $('#custom_snippet_priority').val();
-                var custom_vcl = $('#custom_snippet_content').val();
-                var msgWarning = $('.fastly-message-error');
+                let custom_name = $('#custom_snippet_name').val();
+                let custom_type = $('#custom_snippet_type').val();
+                let custom_priority = $('#custom_snippet_priority').val();
+                let custom_vcl = $('#custom_snippet_content').val();
+                let msgWarning = $('.fastly-message-error');
 
                 if (!custom_name || !custom_type || !custom_priority || !custom_vcl) {
                     msgWarning.text($.mage.__('Please fill out the required fields.')).show();
@@ -1512,7 +1574,8 @@ define([
                         'name': custom_name,
                         'type': custom_type,
                         'priority': custom_priority,
-                        'vcl': custom_vcl
+                        'vcl': custom_vcl,
+                        'edit': false
                     },
                     showLoader: true,
                     success: function (response) {
@@ -1525,6 +1588,59 @@ define([
                                 if (snippetsResp.status != false) {
                                     if (snippetsResp.snippets.length > 0) {
                                         snippets = snippetsResp.snippets;
+                                        vcl.processCustomSnippets(snippets);
+                                    } else {
+                                        $('.no-snippets').show();
+                                    }
+                                }
+                            }).fail(function () {
+                                // TO DO: implement
+                            });
+                        } else {
+                            msgWarning.text($.mage.__(response.msg)).show();
+                        }
+                    },
+                    error: function (msg) {
+                        return errorCustomSnippetBtnMsg.text($.mage.__('An error occurred while processing your request. Please try again.')).show();
+                    }
+                });
+            },
+
+            updateCustomSnippet: function () {
+                let custom_name = $('#custom_snippet_name').val();
+                let original_name = $('#original_snippet_name').val();
+                let custom_type = $('#custom_snippet_type').val();
+                let custom_priority = $('#custom_snippet_priority').val();
+                let custom_vcl = $('#custom_snippet_content').val();
+                let msgWarning = $('.fastly-message-error');
+
+                if (!custom_name || !custom_type || !custom_priority || !custom_vcl) {
+                    msgWarning.text($.mage.__('Please fill out the required fields.')).show();
+                    return;
+                }
+
+                $.ajax({
+                    type: "POST",
+                    url: config.createCustomSnippetUrl,
+                    data: {
+                        'name': custom_name,
+                        'type': custom_type,
+                        'priority': custom_priority,
+                        'vcl': custom_vcl,
+                        'edit': true,
+                        'original': original_name
+                    },
+                    showLoader: true,
+                    success: function (response) {
+                        if (response.status === true) {
+                            active_version = response.active_version;
+                            vcl.modal.modal('closeModal');
+                            successCustomSnippetBtnMsg.text($.mage.__('Custom snippet successfully updated.')).show();
+                            vcl.getCustomSnippets(false).done(function (snippetsResp) {
+                                $('.loading-snippets').hide();
+                                if (snippetsResp.status !== false) {
+                                    if (snippetsResp.snippets.length > 0) {
+                                        let snippets = snippetsResp.snippets;
                                         vcl.processCustomSnippets(snippets);
                                     } else {
                                         $('.no-snippets').show();
@@ -2314,6 +2430,15 @@ define([
                     },
                     actionOk: function () {
                         vcl.setCustomSnippet();
+                    }
+                },
+                'fastly-custom-snippet-edit-options': {
+                    title: jQuery.mage.__('You are about to edit a custom snippet '),
+                    content: function () {
+                        return document.getElementById('fastly-custom-snippet-edit-template').textContent;
+                    },
+                    actionOk: function () {
+                        vcl.updateCustomSnippet();
                     }
                 },
                 'fastly-tls-options': {